Refactor TDDCyclesController and getCommitsUseCase
The commit includes refactoring TDDCyclesController and
getCommitsUseCase to use the correct repositories.
