+++ server/src/routes/TDDCyclesRoutes.ts
@@ -1,13 +1,13 @@
 import express from "express";
-import CommitsController from "../controllers/Commits/commitsController";
 import TDDCyclesController from "../controllers/TDDCycles/TDDCyclesController";
 
 
 const TDDCycles = express.Router();
+const commitsRepository = new CommitsRepository();
+const jobsRepository = new JobsRepository();
+const TDDCyclesController = new TDDCyclesController(commitsRepository, jobsRepository)
 
-const TDDCyclesController = new TDDCyclesController()
-
-TDDCycles.get("/", async(req,res)=> await commitsController.getCommits(req,res));
+TDDCycles.get("/", async (req, res) => await TDDCyclesController.getCommits(req, res));
